Digest errors in cron monitor (#18)
This PR updates the cron monitoring Lambda code so that if a job runs
multiple times during the monitoring window (typically 1 hour) and
reports errors during several of those runs, the monitor rolls the
errors up into one report. This will cut down on the number of messages
sent via SNS (and reported in Slack).

Example output:

```
Message: Fleet cron 'usage_statistics' (last updated 2025-03-24 17:42:05 +0000 UTC) raised errors in 4 of the previous 14 runs; the most recent is:
["4d33d9dd"]
```

(the `["4d33d9dd"]` is a random value added to a cron_stats row for
testing purposes)